// TimeWarden.Client/timewarden-client/src/api/generated/models/invoiceCreateModel.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * TimeWarden.API | v1
 * OpenAPI spec version: 1.0.0
 */
import type { ItemOfWorkCreateModel } from './itemOfWorkCreateModel';

export interface InvoiceCreateModel {
  /** @nullable */
  id?: string | null;
  clientId?: string;
  invoiceDate?: string;
  /** @nullable */
  extraNotes?: string | null;
  itemsOfWork?: ItemOfWorkCreateModel[];
}