// TimeWarden.Client/timewarden-client/src/api/generated/models/itemOfWorkCreateModel.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * TimeWarden.API | v1
 * OpenAPI spec version: 1.0.0
 */

export interface ItemOfWorkCreateModel {
  /** @nullable */
  id?: string | null;
  /** @nullable */
  invoiceId?: string | null;
  /** @nullable */
  projectId?: string | null;
  description?: string;
  /** @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?$ */
  hourlyRate?: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?$ */
  hoursOfWork?: number | string;
  dateOfWork?: string;
}